+++ app/assets/javascripts/administrate/components/table.js
@@ -2,12 +2,14 @@ $(function() {
   var keycodes = { space: 32, enter: 13 };
 
   var visitDataUrl = function(event) {
-    if (event.type=="click" ||
+    if (event.type == "click" ||
         event.keyCode == keycodes.space ||
         event.keyCode == keycodes.enter) {
 
-      if(!event.target.href) {
-        window.location = $(event.target).closest("tr").data("url");
+      if (!event.target.href) {
+        var url = $(event.target).closest("tr").data("url");
+
+        if (url) { window.location = url; }
       }
     }
   };