+++ app/helpers/administrate/application_helper.rb
@@ -1,5 +1,7 @@
 module Administrate
   module ApplicationHelper
+    PLURAL_MANY_COUNT = 2.1
+
     def render_field(field, locals = {})
       locals.merge!(field: field)
       render locals: locals, partial: field.to_partial_path
@@ -12,7 +14,7 @@ module Administrate
         constantize.
         model_name.
         human(
-          count: 0,
+          count: PLURAL_MANY_COUNT,
           default: resource_name.to_s.pluralize.titleize,
         )
     end
@@ -20,17 +22,27 @@ module Administrate
     def svg_tag(asset, svg_id, options = {})
       svg_attributes = {
         "xlink:href".freeze => "#{asset_url(asset)}##{svg_id}",
-        height: options[:height],
-        width: options[:width],
-      }.delete_if { |_key, value| value.nil? }
+        height: "100%",
+        width: "100%",
+      }
       xml_attributes = {
         "xmlns".freeze => "http://www.w3.org/2000/svg".freeze,
         "xmlns:xlink".freeze => "http://www.w3.org/1999/xlink".freeze,
-      }
+        height: options[:height],
+        width: options[:width],
+      }.delete_if { |_key, value| value.nil? }
 
       content_tag :svg, xml_attributes do
         content_tag :use, nil, svg_attributes
       end
+    end
+
+    def sanitized_order_params
+      params.permit(:search, :id, :order, :page, :per_page, :direction, :orders)
+    end
+
+    def clear_search_params
+      params.except(:search, :page).permit(:order, :direction, :per_page)
     end
   end
 end