+++ lib/administrate/engine.rb
@@ -1,3 +1,4 @@
+require "bourbon"
 require "datetime_picker_rails"
 require "jquery-rails"
 require "kaminari"
@@ -20,6 +21,28 @@ module Administrate
   class Engine < ::Rails::Engine
     isolate_namespace Administrate
 
+    @@javascripts = []
+    @@stylesheets = []
+
     Engine.config.assets.precompile << /\.(?:svg)\z/
+
+    def self.add_javascript(script)
+      @@javascripts << script
+    end
+
+    def self.add_stylesheet(stylesheet)
+      @@stylesheets << stylesheet
+    end
+
+    def self.stylesheets
+      @@stylesheets
+    end
+
+    def self.javascripts
+      @@javascripts
+    end
+
+    add_javascript "administrate/application"
+    add_stylesheet "administrate/application"
   end
 end