+++ lib/administrate/field/image.rb
@@ -1,8 +1,0 @@
-require_relative "base"
-
-module Administrate
-  module Field
-    class Image < Field::Base
-    end
-  end
-end