+++ spec/example_app/app/dashboards/customer_dashboard.rb
@@ -7,13 +7,19 @@ class CustomerDashboard < Administrate::BaseDashboard
     email_subscriber: Field::Boolean,
     lifetime_value: Field::Number.with_options(prefix: "$", decimals: 2),
     name: Field::String,
-    orders: Field::HasMany,
+    orders: Field::HasMany.with_options(limit: 2),
     updated_at: Field::DateTime,
+    kind: Field::Select.with_options(collection: Customer::KINDS),
   }
 
   COLLECTION_ATTRIBUTES = ATTRIBUTE_TYPES.keys
   SHOW_PAGE_ATTRIBUTES = ATTRIBUTE_TYPES.keys - [:name]
-  FORM_ATTRIBUTES = [:name, :email, :email_subscriber]
+  FORM_ATTRIBUTES = [
+    :name,
+    :email,
+    :email_subscriber,
+    :kind,
+  ].freeze
 
   def display_resource(customer)
     customer.name