+++ spec/example_app/app/models/customer.rb
@@ -1,8 +1,13 @@
 class Customer < ActiveRecord::Base
-  has_many :orders
+  has_many :orders, dependent: :destroy
 
   validates :name, presence: true
   validates :email, presence: true
+
+  KINDS = [
+    :standard,
+    :vip,
+  ].freeze
 
   def lifetime_value
     orders.map(&:total_price).reduce(0, :+)