+++ spec/features/orders_index_spec.rb
@@ -33,7 +33,7 @@ feature "order index page" do
     order = create(:order)
 
     visit admin_orders_path
-    click_on "Edit"
+    click_on t("administrate.actions.edit")
 
     expect(current_path).to eq(edit_admin_order_path(order))
   end