+++ spec/rails_helper.rb
@@ -7,7 +7,6 @@ require File.expand_path("../../spec/example_app/config/environment", __FILE__)
 require "rspec/rails"
 require "shoulda/matchers"
 require "capybara/poltergeist"
-require "percy/capybara/rspec"
 
 Dir[Rails.root.join("../../spec/support/**/*.rb")].each { |file| require file }
 
@@ -21,6 +20,7 @@ end
 RSpec.configure do |config|
   config.include Features, type: :feature
   config.include DashboardHelpers
+  config.include ControllerHelpers
   config.infer_base_class_for_anonymous_controllers = false
   config.infer_spec_type_from_file_location!
   config.use_transactional_fixtures = false